Trait pallet_bonded_coins::curves::BondingFunction

pub trait BondingFunction<Balance> {
    // Required method
    fn calculate_costs(
        &self,
        low: Balance,
        high: Balance,
        passive_supply: Vec<Balance>
    ) -> Result<Balance, ArithmeticError>;
}

Trait defining the bonding function for a curve. The bonding function is used to calculate the cost of purchasing or selling assets using the curve. The trait is implemented for each curve type.

Parameters:

low: The lower bound of integral.

high: The upper bound of integral.

passive_supply: The passive supply of other assets in the pool, which are not affected by the operation.

Required Methods

fn calculate_costs( &self, low: Balance, high: Balance, passive_supply: Vec<Balance> ) -> Result<Balance, ArithmeticError>