Function pallet_bonded_coins::pallet::dispatchables::create_pool

pub fn create_pool<T: Config>(
    curve: CurveInput<<T as Config>::CurveParameterInput>,
    collateral_id: CollateralAssetIdOf<T>,
    currencies: BoundedVec<TokenMeta<<<T as Config>::Fungibles as InspectFungibles<<T as Config>::AccountId>>::Balance, BoundedVec<u8, <T as Config>::MaxStringInputLength>, BoundedVec<u8, <T as Config>::MaxStringInputLength>>, T::MaxCurrenciesPerPool>,
    denomination: u8,
    transferable: bool,
    min_operation_balance: u128
)

Creates a new bonded token pool. The pool will be created with the given curve, collateral currency, and bonded currencies. The pool will be owned by the origin account.

Parameters

origin: The origin of the call.

curve: The curve parameters for the pool.

collateral_id: The ID of the collateral currency.

currencies: A bounded vector of token metadata for the bonded currencies.

denomination: The denomination for the bonded currencies.

transferable: A boolean indicating if the bonded currencies are transferable.

Returns

DispatchResult: The result of the dispatch.

Errors

Error::<T>::InvalidInput: If the denomination is greater than the maximum allowed or if the curve input is invalid.

Error::<T>::Internal: If the conversion to BoundedVec fails.

Other errors depending on the types in the config.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::create_pool.