Function pallet_bonded_coins::pallet::dispatchables::mint_into

pub fn mint_into<T: Config>(
    pool_id: T::PoolId,
    currency_idx: u32,
    beneficiary: <<T as Config>::Lookup as StaticLookup>::Source,
    amount_to_mint: <<T as Config>::Fungibles as InspectFungibles<<T as Config>::AccountId>>::Balance,
    max_cost: <<T as Config>::Collaterals as InspectFungibles<<T as Config>::AccountId>>::Balance,
    currency_count: u32
)

Mints new bonded tokens. The tokens will be minted into the beneficiary account. In exchange, an amount of collateral determined by the pool’s bonding curve is debited from the caller and transferred to the pool account. The origin account must be a manager of the pool if its state is Locked and allow_mint is false. The pool must be in a live (non-refunding, non-destroying) state.

Parameters

origin: The origin of the call.

pool_id: The identifier of the pool.

currency_idx: The index of the currency in the bonded currencies vector.

beneficiary: The account to receive the minted tokens.

amount_to_mint: The amount of bonded tokens to mint.

max_cost: The maximum cost of collateral.

currency_count: The maximum number of currencies allowed in the pool.

Returns

DispatchResultWithPostInfo: The result of the dispatch with the actual used weights.

Errors

Error::<T>::PoolUnknown: If the pool does not exist.

Error::<T>::NoPermission: If the caller does not have permission to mint.

Error::<T>::CurrencyCount: If the number of currencies exceeds currency_count.

Error::<T>::IndexOutOfBounds: If the currency index is out of bounds.

ArithmeticError::Overflow: If there is an overflow during the calculation.

Error::<T>::Slippage: If the cost exceeds max_cost.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::mint_into.