Trait pallet_bonded_coins::pallet::Config

pub trait Config: Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type DepositCurrency: MutateHold<Self::AccountId, Reason = Self::RuntimeHoldReason>;
    type Collaterals: MutateFungibles<Self::AccountId> + AccountTouch<CollateralAssetIdOf<Self>, Self::AccountId> + InspectMetadata<Self::AccountId>;
    type Fungibles: CreateFungibles<Self::AccountId> + DestroyFungibles<Self::AccountId> + MutateMetadata<Self::AccountId> + InspectMetadata<Self::AccountId> + MutateFungibles<Self::AccountId> + FreezeAccounts<Self::AccountId, FungiblesAssetIdOf<Self>> + ResetTeam<Self::AccountId>;
    type MaxCurrenciesPerPool: Get<u32>;
    type MaxStringInputLength: Get<u32>;
    type MaxDenomination: Get<u8>;
    type DepositPerCurrency: Get<<<Self as Config>::DepositCurrency as InspectFungible<<Self as Config>::AccountId>>::Balance>;
    type BaseDeposit: Get<<<Self as Config>::DepositCurrency as InspectFungible<<Self as Config>::AccountId>>::Balance>;
    type DefaultOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>;
    type PoolCreateOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>;
    type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type PoolId: Parameter + MaxEncodedLen + From<[u8; 32]> + Into<Self::AccountId>;
    type RuntimeHoldReason: From<Self::HoldReason>;
    type HoldReason: TryFrom<Self::PoolId>;
    type CurveParameterType: Parameter + Member + FixedSigned + MaxEncodedLen + PartialOrd<I9F23> + From<I9F23>;
    type CurveParameterInput: Parameter + FixedUnsigned + MaxEncodedLen;
    type WeightInfo: WeightInfo;
    type NextAssetIds: NextAssetIds<Self>;
    type BenchmarkHelper: BenchmarkHelper<Self>;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type DepositCurrency: MutateHold<Self::AccountId, Reason = Self::RuntimeHoldReason>

The currency used for storage deposits.

type Collaterals: MutateFungibles<Self::AccountId> + AccountTouch<CollateralAssetIdOf<Self>, Self::AccountId> + InspectMetadata<Self::AccountId>

A fungibles trait implementation to interact with currencies which can be used as collateral for minting bonded tokens.

type Fungibles: CreateFungibles<Self::AccountId> + DestroyFungibles<Self::AccountId> + MutateMetadata<Self::AccountId> + InspectMetadata<Self::AccountId> + MutateFungibles<Self::AccountId> + FreezeAccounts<Self::AccountId, FungiblesAssetIdOf<Self>> + ResetTeam<Self::AccountId>

Implementation of creating and managing new fungibles

type MaxCurrenciesPerPool: Get<u32>

The maximum number of currencies allowed for a single pool.

type MaxStringInputLength: Get<u32>

type MaxDenomination: Get<u8>

The maximum denomination that bonded currencies can use. This should be configured so that 10^MaxDenomination < 2^CurveParameterType::frac_nbits() as larger denominations could result in truncation.

type DepositPerCurrency: Get<<<Self as Config>::DepositCurrency as InspectFungible<<Self as Config>::AccountId>>::Balance>

The deposit required for each bonded currency.

type BaseDeposit: Get<<<Self as Config>::DepositCurrency as InspectFungible<<Self as Config>::AccountId>>::Balance>

The base deposit required to create a new pool.

type DefaultOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>

The origin for most permissionless and privileged operations.

type PoolCreateOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>

The dedicated origin for creating new bonded currency pools (typically permissionless).

type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin for permissioned operations (force_* transactions).

type PoolId: Parameter + MaxEncodedLen + From<[u8; 32]> + Into<Self::AccountId>

The type used for pool ids

type RuntimeHoldReason: From<Self::HoldReason>

type HoldReason: TryFrom<Self::PoolId>

type CurveParameterType: Parameter + Member + FixedSigned + MaxEncodedLen + PartialOrd<I9F23> + From<I9F23>

The type used for the curve parameters. This is the type used in the calculation steps and stored in the pool details.

type CurveParameterInput: Parameter + FixedUnsigned + MaxEncodedLen

Input type for curve parameters. This is the type used in the extrinsic calls to create a new pool. It is converted to CurveParameterType.

type WeightInfo: WeightInfo

type NextAssetIds: NextAssetIds<Self>

type BenchmarkHelper: BenchmarkHelper<Self>

Benchmark helper to calculate asset ids for the collateral and bonded currencies.