Type Alias pallet_bonded_coins::pallet::PoolDetailsOf

pub type PoolDetailsOf<T> = PoolDetails<<T as Config>::AccountId, Curve<<T as Config>::CurveParameterType>, BoundedVec<FungiblesAssetIdOf<T>, <T as Config>::MaxCurrenciesPerPool>, CollateralAssetIdOf<T>, <<T as Config>::DepositCurrency as InspectFungible<<T as Config>::AccountId>>::Balance>;

Aliased Type

struct PoolDetailsOf<T> {
    pub owner: <T as Config>::AccountId,
    pub manager: Option<<T as Config>::AccountId>,
    pub curve: Curve<<T as Config>::CurveParameterType>,
    pub collateral: <<T as Config>::Collaterals as Inspect<<T as Config>::AccountId>>::AssetId,
    pub bonded_currencies: BoundedVec<<<T as Config>::Fungibles as Inspect<<T as Config>::AccountId>>::AssetId, <T as Config>::MaxCurrenciesPerPool>,
    pub state: PoolStatus<Locks>,
    pub transferable: bool,
    pub denomination: u8,
    pub min_operation_balance: u128,
    pub deposit: <<T as Config>::DepositCurrency as Inspect<<T as Config>::AccountId>>::Balance,
}

Fields
owner: <T as Config>::AccountId

The owner of the pool.

manager: Option<<T as Config>::AccountId>

The manager of the pool. If a manager is set, the pool is permissioned.

curve: Curve<<T as Config>::CurveParameterType>

The curve of the pool.

collateral: <<T as Config>::Collaterals as Inspect<<T as Config>::AccountId>>::AssetId

The collateral currency of the pool.

bonded_currencies: BoundedVec<<<T as Config>::Fungibles as Inspect<<T as Config>::AccountId>>::AssetId, <T as Config>::MaxCurrenciesPerPool>

The bonded currencies of the pool.

state: PoolStatus<Locks>

The status of the pool.

transferable: bool

Whether the pool is transferable or not.

denomination: u8

The denomination of the pool.

min_operation_balance: u128

The minimum amount that can be minted/burnt.

deposit: <<T as Config>::DepositCurrency as Inspect<<T as Config>::AccountId>>::Balance

The deposit to be returned upon destruction of this pool.