Trait pallet_bonded_coins::traits::ResetTeam

pub trait ResetTeam<AccountId>: Inspect<AccountId> {
    // Required method
    fn reset_team(
        id: Self::AssetId,
        owner: AccountId,
        admin: AccountId,
        issuer: AccountId,
        freezer: AccountId
    ) -> DispatchResult;
}

Copy of a trait from a later version of the Polkadot SDK (frame_support::traits::tokens::fungibles::roles::ResetTeam). Once we upgraded to Polkadot SDK version 1.13.0, this can be retired in favor of the original trait.

Required Methods

fn reset_team( id: Self::AssetId, owner: AccountId, admin: AccountId, issuer: AccountId, freezer: AccountId ) -> DispatchResult

Reset the team for the asset with the given id.

Parameters

id: The identifier of the asset for which the team is being reset.

owner: The new owner account for the asset.

admin: The new admin account for the asset.

issuer: The new issuer account for the asset.

freezer: The new freezer account for the asset.